Fixed tzfile parsing for leapcnt-containing tzfile, with tests.
The previous implementation contained a commented code part which must not be
omitted if leapcnt is set or otherwise the following parts read from the file are
actually garbled, so this code has been uncommented. The existing
testLeapCountDecodesProperly test didn't really cover this at all which has been
fixed. Additional test methods for appropriate reading of the isstd and isgmt
information contained in the TZTest.NEW_YORK tzfile data have been added.