Fix warning on French Windows systems when parsing date times under Python 2
When parsing a date time in some conditions,
``dateutil.parser.parse()`` emits the following warning:

   UnicodeWarning: Unicode equal comparison failed to convert both arguments to Unicode - interpreting them as being unequal

This warning comes when testing if a unicode timezone name is in
``time.tzname``.  I'm not sure exactly how to trigger this code
path, but I can reproduce the warning on my system using the
following snippet:

   import time
   print time.tzname
   print u'EST' in time.tzname

The problem is that ``time.tzname[1]`` contains a byte string with
non-ASCII characters in the current code page.

To remove the warning, we should ensure to decode ``time.tzname``
using ``locale.getpreferredencoding()`` so that we have somthing
that can safely compare with unicode strings.