Fix bug on French Windows systems when using local timezone under Python 2
When I try to run the following snippet, I get the following output:

    >>> from datetime import datetime
    >>> from dateutil.tz import tzlocal
    >>>
    >>> now = datetime.now()
    >>> print tzlocal().tzname(now)

    Traceback (most recent call last):
      File "<stdin>", line 1, in <module>
      File "...\dateutil\tz\_common.py", line 22, in adjust_encoding
        name = name.encode()
    UnicodeDecodeError: 'ascii' codec can't decode byte 0x92 in position 12: ordinal not in range(128)

This happens because ``time.tzname[1]`` contains non-ASCII
characters in the current code page.  We need to decode this
using ``locale.getpreferredencoding()`` to get a Unicode string
that is safe to use.

Additionally, since Python 2 forces ``.tzname()`` to return a
byte string, the ``adjust_encoding()`` utility tries to re-encode
to ASCII.  This is obviously not possible so the best path seems
to be to remove diacritics to lossily convert the stirng to ASCII.