Change weakref test mechanism
This is a more reliable mechanism to test that nothing continues
to hold a strong reference to the cached values.

I have also switched the tests over to using pytest-style bare
functions, and marked them all as smoke tests, since this will
not actually be a guaranteed behavior of the interface.