Drop support for rscale in rrule
Since we do not actually have support for RSCALE in the rrule class,
it is premature to allow users to specify it. Instead, we will validate
it if present in strings and emit it if necessary to comply with the RFC
but otherwise postpone an explicit API for handling RSCALE until after
real support is added.