Add bounds to dst offset in tzrange constructor
This adds the same bounds-checking to the DST offset construction that
is applied to the STD offset. It's slightly more complicated because DST
may be implicitly set as a function of STD.

This commit also tweaks the way STD offset bounds are checked to avoid
unnecessary runtime calculation of the number of seconds in 24 hours.