Move UTC definition into dateutil.tz.tz
To allow the UTC object to be accessible from the functions in
`dateutil.tz.tz` such as `gettz` and datetime_exists, it is necessary to
define the UTC constant here and import it into `dateutil.tz`. This is a
very minor performance improvement, avoiding a call to the `tzutc()`
constructor.