Fix incorrect import in Python 3.7+
It turns out that the import of `contextlib.nullcontext` in Python 3.7+
was accidentally trying to import `contextmanager.nullcontext`, but this
was never noticed because it simply fell back to the backport included
here.

I do not know of a great way to test this that doesn't rely on
implementation details, but I have manually tested it by setting a
breakpoint after the import in the `try` statement and importing
`dateutil.tz` manually on Python 3.7, so this version should work.