Fix Travis test for build command.
The orignal test was using `[ ! -e "dist/*.whl" ]` to try and detect
that a wheel was created (and an analogous command for the tarball), but it
turns out that this would succeed even if the build command completely fails!
The reason is that the shellw as expanding "dist/*.whl" to an empty
list, which the conditional took to mean that it didn't need to check
for the existence of *any* files.

This was replaced by a more robust command that checks that exactly one
of each file was created by the build.