Skip coverage uploading for Python 3.4 on Appveyor
Coverage 5.x has made some backwards-incompatible changes to its format
and cannot understand the format generated by coverage 4.x. Since
coverage 5.x does not support Python 3.4 and the "combine and upload"
job is using Python 3.6 (and thus `coverage >= 5`), coverage uploading
was broken.

It is possible to get this working again, but probably more trouble than
it is worth, since as far as I can tell there are no lines hit *only* by
the Python 3.4 tests on Windows, so we can leave it out of the
combined coverage measurement.