Decode input str when reporting an error in `parse_isodate`
As the function is wrapped around a decorator that converts the input to
bytes, when an exception is raised that includes the value, it contains
the value as bytes (b"<val>"). To provide the expected str in Python 3,
decode that to unicode.

This will result in `u""` for Python 2 callers, but given that we plan
to drop support for Python 2 in the short term, it feels acceptable.