ci: Upgrade setuptools when installing dependencies
We are using `setup_requires` which uses `easy_install` in old versions
of `setuptools` rather than pip. This results in the installation of the
latest version of `setuptools_scm`, as `easy_install` does not support
`python_requires` tags. This commit runs the CI with the latest version
available of `setuptools` for Python 3.4, which should be frozen.