Remove test of implementation details
This tests implementation details of `tzfile` which are subject to
change. The values read for `isstd` only really matter insofar as they
affect the results of `utcoffset()` calls, which should be well covered
by other tests. Given that we are going to change these implementation
details anyway, it's time to remove this test.