Update information about release signatures
The industry is moving away from GPG signatures, and signing
distribution artifacts will complicate the release process and make it
harder to allow other maintainers to make releases for only minor gain
in auditability of releases.

At least for now, we'll remove any guarantees about signatures and trust
GitHub's authentication.

Note: This commit is signed with the same key that signed releases
2.4.1-2.8.2.