Implement ChainParser
Add a class that will provide users with a new parser type that allows
parsing a string using a list of parsers or string formats to be used in
strptime.

This class can be used when users have an idea of potential formats that
the string might have been formatted with and still allows to be
combined with other dateutils parsers.