Fix test failure in 2027
Santiago Vila pointed out in https://bugs.debian.org/1078638 that one of
the test cases in `tests/test_parser.py::test_parser` will start failing
soon, specifically at the start of 2027; this is because two-digit years
are converted to a year within [-50, 49] of the current year.

To avoid this, run tests of parsing two-digit years with a frozen time.