Fix dead link in datetime.easter.easter documentation
The [previous link](http://www.gmarts.org/index.php?go=415) resolves
and does not produce an error, which is why it was not caught by the
documentation checks in CI.  However, the content displayed is now a
generic message indicating that the `gmarts.org` website has been shut
down.

The [updated
link](https://web.archive.org/web/20241001041454/http://dates.gmarts.org/eastalg.htm)
is to the `archive.org` mirror of the page, for the last recorded
snapshot in October 2024.